CMake: Runtimes: Specify language mode explicitly.
The compiler will soon require that any module with a `.swiftinterface` be
built with an explicit language mode specified on the command line. This
prevents misinterpretation of the `.swiftinterface` when building it for
clients that aren't building with the default language mode.

Add explicit `-swift-version` arguments to the builds of various Runtimes
libraries to comply with this new requirement.