[DiagnoseStaticExclusivity] Fix this assertion.
This function is called in one place in an assertion.  It asserts that
the function in question is the stdlib's swap.  In language mode 6, the
called value may be wrapped in a `function_conversion_expr` for
`@Sendable`.  That's irrelevant for the purposes of this assertion, so
look through such conversions.

rdar://160692694