build.ps1: stage static library artifacts
These dependencies are normally consumed by the dynamic target that uses
them. However, when distributing a static SDK, we must distribute the
static library content as well to allow linking against the version of
the dependency that the runtime is built against. The binary staging
here is required to allow us to package the content into the toolchain
distribution.