[sil] Add a new instruction inject_guaranteed.
This instruction is intended to be used to convert non-Trivial values with none/unowned ownership
into guaranteed values that are sub-values of a base value. It is intended to be
used in situations where one has taken a guaranteed value, cast the guaranteed
value to a trivial form using bitcasts, manipulated the guaranteed value, and
then cast it back. An example of such a situation is when masking out the low
bits of a tagged pointer. E.x.:

```
bb0(%0 : @guaranteed $Class):
  %1 = unchecked_value_cast %0 : $Class to $Builtin.Word
  %2 = integer_literal $Builtin.Word, -7
  %3 = buildin "and_Word"(%1, %2)
  // %4 has none ownership
  %4 = unchecked_value_cast %3 : $Builtin.Word to $Class
  %5 = inject_guaranteed %4 : $Class onto %0 : $Class
  // Use %5. %5 does not need an end_borrow since it is considered
  // to be part of %0's borrow scope.
```

We could use unchecked_ownership_conversion for this type of thing but the
resulting value would be viewed as a new independent guaranteed value (sort of
like a guaranteed function argument) rather than as a guaranteed-forwarding
projection out of a different guaranteed value.