Rename the new `_GUID` clang module to `_GUIDDef`.
This PR changes the name of the clang module introduced in #83914 from `_GUID`
to `_GUIDDef`. This change is necessary because the Windows SDK declares a type
named `_GUID` and Swift gets confused when a module and a type have the same
name.

This confusion is the motivation for [SE-0491](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0491-module-selectors.md),
but we can't just wait for a fix because the new module is a source-breaking
change and clients won't be able to adopt the proposed syntax for some time.