[CoroutineAccessors] Fix overriding.
Before the feature is enabled, a var decl for which a reading accessor
is implemented via _read still had as its opaque accessor a getter
(unless it's marked @_borrowed, it's noncopyable, or it has a lifetime
dependency).  So when that `_read` is replaced with a `read`/`yielding
borrow`, it is still necessary to emit a getter.  And for back
deployment it is still necessary to provide a default override of the
getter in terms of the `read`.

rdar://149352519