[cxx-interop] Fix inherited nested types (#84455)
When a derived class inherits a nested type from a base class in C++, we
expect that derived type member to be the same as the base nested type.
This patch fixes the ClangImporter's base member cloning to make that
the case, and also fixes up some test cases (renames "sub-type" to
"nested type" because the term "subtype" is overloaded and misleading
here).

rdar://161137420