OpenBSD 7.8 will require OpaquePointer for FILE.
OpenBSD 7.8 snapshots hide the type information for FILE. Therefore, the
types for the standard stdio streams should (regrettably) be
OpaquePointer, due to the well-discussed issue of losing type information
for forward-declared C types.

We explicitly drop to void * in the LibcOverlayShims for backwards
compatibility, since OpenBSD 7.8 is not yet released.

However, because SwiftReflectionTest uses libc FILE functions like
fwrite and fflush directly, this means that building this on 7.7 will
still fail because types will mismatch. Instead of adding extra shims,
let's just skip building SwiftReflectionTest on 7.7.