[Debug] Add pointer based stringForPrintObject
Adds an overload of `stringForPrintObject` which takes a pointer and mangled typename as
arguments. This will be used to improve performance and resilience of `po` in lldb.

The pointer and mangled typename are used to construct an `Any` value, which is then
passed into the primary implementation of `stringForPrintObject`.

This allows calling `stringForPrintObject` without having to first construct a context
that contains all necessary Swift modules. This will improve speed, and also resilience
when modules cannot be loaded for whatever reason.