Add an _isNFDQuickCheck property to scalars
This adds a property for the derived "NFD_QC" property. The naming /
implementation is a little wonky, since the scalars shown in the
derived properties file are those for which the property is _false_.

The purpose of this is to identify scalars that don't have any other
canonically equivalent sequences, which lets us search directly for
their byte sequence rather than needing to normalize a string while
searching.