Sema: Workaround for broken existential opening behavior
This is a narrow workaround for a regression from
9e3d0e0a8c78076a70a9869d10acc6189cb7bb6b.

There is no reason to skip this logic for protocol extension
members, except that doing so happens to break existential
opening in an expression that involves a call to Array.init
elsewhere.

However there is an underlying issue here with existential opening,
which doesn't seem to work right in the presence of overloading.

The test case demonstrates the fixed problem, together with an
existing bug that points to the underlying problem.

Fixes rdar://160389221.