CMake: Set language version across supplemental libs
The Swift language version appears in the Swift interface and impacts
how the interface is handled by the compiler when importing the module.
The version only appears if it is set, which means that downstream
projects setting the version explicitly may interpret the swift
interface file differently.

This also factors the setting out into a separate file, which is used to
set the C++ standard and Swift language version since it is the same
across these libraries at this time. If this assumption changes, we can
refactor as necessary and this structure shouldn't be seen as
enforcement.

Finally, this patch moves setting the language version and standard
variables before the languages are enabled. This ensures that the
try-compile applies the appropriate flags when verifying that the
compiler works.