Turn swift_asyncLet_finish into a builtin.
Reflect the potential stack allocation of the begin/finish pair in SIL,
which should prevent the optimizer from trying to move stack allocations
around them.

I'm not actually sold that the runtime *has* to do a stack allocation here,
because it only really does it because it creates a result buffer in the
future that it should be able to avoid. But it's undisputable that it
currently does so, and we may come up with good reasons for it to continue
to do so even if we remove the current allocation, so modeling the allocation
is the right thing to do.

Fixes rdar://109850951