Provide an implementation of `swift_stdlib_random()` derived from `std::random_device`.
This PR provides an alternate platform-agnostic implementation of
`swift_stdlib_random()` based on C++11's `std::random_device`. This
implementation should be portable even to Embedded Swift, with the significant
caveat that there is no way to know at compile time if the random numbers
produced are _actually_ non-deterministic. (They _should_ be, but if the current
implementation has no suitable RNG, the API doesn't provide a way to detect that
statically.)