[test] Fix compatibility span calculation for unified builds
In unified builds `swift_obj_root` is in a different place in the CMake
build tree, so using it as the base for finding `lib` is not correct.

Change the usage of `swift_obj_root/lib` for `config.swift_lib_dir`
which should be correct in both unified and non-unified builds.