[ast] Requestify performImportResolution.
When making improvements to sil_llvm_opt_main, I noticed that when we performed
IRRequest, we hadn't performed Sema... so I then added code that ran Sema. I
then discovered that OptimizedIRRequest actually also performed Sema causing a
crash to occur. It then appeared to me... no one had requestified
performImportResolution even though it is a perfect case. I am not going to need
to do this soon (I am going to move a performImportResolution into IRRequest)...
but it seemed like a good thing to do and a clear improvement.