[test] Mark test as UNSUPPORTED when maccatalyst_support is enabled
If one tries to enable maccatalyst_support (which doesn't work
correctly, see #77506), this test fails because `Swift.swiftmodule` gets
generated with both `-target` and `-target-variant`
`swift-module-flags`, which triggers the usage of the
`isOSVersionAtLeastOrVariantAtLeast` version of version check, which
does not match the test `CHECK` lines.

Mark the test as `UNSUPPORTED` when the feature is enabled. Not sure if
the test can be rewritten conditionally on a feature being support and
checking for different functions in the IR.

This should not affect swift.org CI because that CI never enables
maccatalyst_support.