SILOptimizer: Fix mandatory switch dispatch elimination for unreachable cases.
https://github.com/swiftlang/swift/pull/67920 introduced a mandatory SIL
optimization that eliminates switch statement dispatch for enum elements that
are provably unavailable at runtime. This optimization helps eliminate
unreachable code, but it's also load bearing because it removes references to
enum element symbols that may not preset at link time. The optimization was
checking the wrong condition to determine whether an enum element is
unavailable at runtime and so the optimization wasn't working for code using
custom availability domains.

Resolves rdar://161846311.